Introduce pending `PINFuture`
This commit contains a proposal for introducing a version of `PINFuture`
that waits to be fulfilled or rejected by the user.

This would be useful in cases where we need to wait for a delegate call
to happen, a 3rd party library or SDK calling its delegate on some event
for example, and want to chain calls around the future.